# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel App API",
        "description": "Documentación de la API Travel App",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/carroceriaCoches": {
            "get": {
                "tags": [
                    "Carrocerías de Coche"
                ],
                "summary": "Listar todas las carrocerías de coche",
                "description": "Obtiene la lista de todas las carrocerías de coche disponibles. Solo accesible para usuarios registrados",
                "operationId": "7bf67489fc9501de61583c4fa328c569",
                "responses": {
                    "200": {
                        "description": "Listado de carrocerías de coche",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CarroceriaCoche"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Carrocerías de Coche"
                ],
                "summary": "Crear una nueva carrocería de coche",
                "description": "Solo accesible para administradores y superadministradores. Permite crear y añadir una nueva carrocería a la Base de Datos.",
                "operationId": "4008778e1e661063a30ae7d9adcf0b3d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarroceriaCocheDB"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Carrocería de coche creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CarroceriaCoche"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acceso denegado. Solo administradores pueden crear carrocerías.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Solo los administradores pueden guardar una nueva carrocería"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "nombre": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Ya existe esa carrocería en nuestra BBDD'"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/carroceriaCoches/{id}": {
            "get": {
                "tags": [
                    "Carrocerías de Coche"
                ],
                "summary": "Mostrar una carrocería de coche específica (por id)",
                "description": "Obtiene los detalles de la carrocería de un coche por su ID. Solo accesible para usuarios registrados",
                "operationId": "ab83cd6772a9e87b23f09f48cfaaee34",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la carrocería",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Carrocería encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CarroceriaCoche"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Carrocería no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Carrocería no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Carrocerías de Coche"
                ],
                "summary": "Actualizar una carrocería de coche",
                "description": "Permite modificar los datos de la carrocería de un coche. Solo accesible para administradores y superadministradores.",
                "operationId": "1d44de6dbf5c60a791153e1c3e0feedb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la carrocería a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarroceriaCocheDB"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Carrocería actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CarroceriaCoche"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permisos para actualizar",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Solo el administrador puede editar una carrocería"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Carrocería no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Carrocería no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El campo nombre es obligatorio."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Carrocerías de Coche"
                ],
                "summary": "Borrar una carrocería de coche",
                "description": "Permite eliminar una carrocería de coche. Solo accesible para administradores y superadministradores.",
                "operationId": "b26c48ceed21617556298a2b05abc6c4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la carrocería a borrar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Carrocería borrada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Carrocería borrada correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permisos para borrar",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Solo el administrador puede borrar una carrocería"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Carrocería no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Carrocería no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CarroceriaCoche": {
                "title": "Carrocería de Coche (Recurso API)",
                "description": "Recurso de carrocería de un coche (respuesta)",
                "required": [
                    "id",
                    "nombre"
                ],
                "properties": {
                    "id": {
                        "description": "ID de la carrocería del coche",
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "description": "Nombre de la carrocería del coche",
                        "type": "string",
                        "example": "SUV"
                    }
                },
                "type": "object"
            },
...
            "CarroceriaCocheDB": {
                "title": "Carrocería de Coche (BBDD)",
                "description": "Esquema para crear o actualizar una carrocería de coche (entrada)",
                "required": [
                    "nombre"
                ],
                "properties": {
                    "nombre": {
                        "description": "Nombre de la carrocería",
                        "type": "string",
                        "example": "Hatchback"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}